requirements.txt
fastapi
pydantic
uvicorn[standard]
websockets
mysql-connector-python
pandas
python-dotenv
asyncio
requests
geopy